fix component name for readability
